+++ utoipa-gen/src/component/into_params.rs
@@ -160,7 +160,9 @@ impl ToTokensDiagnostics for IntoParams {
         tokens.extend(quote! {
             impl #impl_generics utoipa::IntoParams for #ident #ty_generics #where_clause {
                 fn into_params(parameter_in_provider: impl Fn() -> Option<utoipa::openapi::path::ParameterIn>) -> Vec<utoipa::openapi::path::Parameter> {
-                    vec![#(#params),*].into_iter().filter(Option::is_some).flatten().collect()
+                    let mut parameters = Vec::<utoipa::openapi::path::Parameter>::new();
+                    #( parameters.extend(#params); )*
+                    parameters
                 }
             }
         });
@@ -325,6 +327,22 @@ impl Param {
     ) -> Result<Self, Diagnostics> {
         let mut tokens = TokenStream::new();
         let field_serde_params = &field_serde_params;
+        let mut field_features = field_features;
+
+        if field_serde_params.flatten {
+            let ignore = pop_feature!(field_features => Feature::Ignore(_));
+            let schema_with =
+                pop_feature!(field_features => Feature::SchemaWith(_) as Option<SchemaWith>);
+            return Ok(Self {
+                tokens: into_params_flatten_tokens(
+                    field,
+                    ignore,
+                    schema_with,
+                    &container_attributes,
+                ),
+            });
+        }
+
         let ident = &field.ident;
         let mut name = &*ident
             .as_ref()
@@ -347,7 +365,7 @@ impl Param {
         let should_always_ignore = matches!(&ignore, Some(Feature::Ignore(Ignore(LitBoolOrExprPath::LitBool(b)))) if b.value());
         if should_always_ignore {
             return Ok(Self {
-                tokens: quote! { None },
+                tokens: quote! { Vec::<utoipa::openapi::path::Parameter>::new() },
             });
         }
         let rename = pop_feature!(param_features => Feature::Rename(_) as Option<Rename>)
@@ -436,9 +454,9 @@ impl Param {
             Some(Feature::Ignore(Ignore(LitBoolOrExprPath::LitBool(bool)))) => {
                 quote_spanned! {
                     bool.span() => if #bool {
-                        None
+                        Vec::<utoipa::openapi::path::Parameter>::new()
                     } else {
-                        Some(#tokens)
+                        vec![#tokens]
                     }
                 }
             }
@@ -448,14 +466,14 @@ impl Param {
                         utoipa::__dev::warn_deprecated_ignore_fn_pattern();
 
                         if #path() {
-                            None
+                            Vec::<utoipa::openapi::path::Parameter>::new()
                         } else {
-                            Some(#tokens)
+                            vec![#tokens]
                         }
                     }
                 }
             }
-            _ => quote! { Some(#tokens) },
+            _ => quote! { vec![#tokens] },
         };
 
         Ok(Self { tokens })
@@ -510,6 +528,57 @@ impl Param {
                 (schema_features, param_features)
             },
         ))
+    }
+}
+
+fn into_params_flatten_tokens(
+    field: &Field,
+    ignore: Option<Feature>,
+    schema_with: Option<SchemaWith>,
+    container_attributes: &FieldParamContainerAttributes<'_>,
+) -> TokenStream {
+    let should_always_ignore = matches!(&ignore, Some(Feature::Ignore(Ignore(LitBoolOrExprPath::LitBool(b)))) if b.value());
+    if should_always_ignore {
+        return quote! { Vec::<utoipa::openapi::path::Parameter>::new() };
+    }
+
+    let field_ty = &field.ty;
+    let parameter_in_provider = match container_attributes.parameter_in {
+        Some(Feature::ParameterIn(parameter_in)) => quote! { || Some(#parameter_in) },
+        _ => quote! { || parameter_in_provider() },
+    };
+    let tokens = if let Some(schema_with) = schema_with {
+        schema_with.to_token_stream_with_args(parameter_in_provider)
+    } else {
+        quote_spanned! {
+            field_ty.span() => <#field_ty as utoipa::IntoParams>::into_params(#parameter_in_provider)
+        }
+    };
+
+    match ignore {
+        Some(Feature::Ignore(Ignore(LitBoolOrExprPath::LitBool(bool)))) => {
+            quote_spanned! {
+                bool.span() => if #bool {
+                    Vec::<utoipa::openapi::path::Parameter>::new()
+                } else {
+                    #tokens
+                }
+            }
+        }
+        Some(Feature::Ignore(Ignore(LitBoolOrExprPath::ExprPath(path)))) => {
+            quote_spanned! {
+                path.span() => {
+                    utoipa::__dev::warn_deprecated_ignore_fn_pattern();
+
+                    if #path() {
+                        Vec::<utoipa::openapi::path::Parameter>::new()
+                    } else {
+                        #tokens
+                    }
+                }
+            }
+        }
+        _ => tokens,
     }
 }